refactor(libs): searchForm default values & validation
