+++ libs/types/index.ts
@@ -1,3 +1,3 @@
 export * from './components';
 export * from './enums';
-
+export * from './validation';
