+++ libs/types/validation/index.ts
@@ -1,0 +1,1 @@
+export * from './searchValidation';
