+++ libs/utils/index.ts
@@ -3,5 +3,5 @@ export * from './generateFactionBackground';
 export * from './generateItemBackground';
 export * from './generateItemTitle';
 export * from './randomInt';
-export * from './submitSearchForm';
 export * from './convertDate';
+export * from './initialValues';
